Refactor API into routers
